LLM server/chunker/requirements.txt
langchain>=0.0.302
python-docx>=0.8.11
nltk>=3.8.1
tiktoken>=0.5.0